requirements.txt
# ipywidgets==8.1.2
# matplotlib==3.8.4
# pandas==2.2.2
# torch==2.8.0
# torchvision==0.23.0
# tqdm==4.66.4

# numpy = 2.2.6
pandas = 2.3.3
python-okx = 0.4.0